+++ CHANGES.rst
@@ -2,6 +2,7 @@
 ====================
 
 - added :class:`udev.Monitor` to support event monitoring
+- added :meth:`udev.Device.asbool` and :meth:`udev.Device.asint`
 - removed type magic in :meth:`udev.Device.__getitem__`
 
 0.1 (May 03, 2010)